Trekkjer batch-mekanismen heilt ut til eiga fil
